+++ src/pages/index.tsx
@@ -1,6 +1,7 @@
 import { Banner } from '@/components/Banner';
 import { Main } from '@/components/Main';
 import { Nav } from '@/components/Nav/Nav';
+
 import { About } from '@/sections/index/About';
 import { Hero } from '@/sections/index/Hero';
 import { Joke } from '@/sections/index/Joke';
