+++ src/sections/index/Technologies.tsx
@@ -1,13 +1,13 @@
-import { Box } from '../../components/Box';
-import { Section } from '../../components/Section';
-import { Inline } from '../../components/Typography/Inline';
-import { SectionTitle } from '../../components/Typography/SectionTitle';
+import { Box } from '@/components/Box';
+import { Section } from '@/components/Section';
+import { Inline } from '@/components/Typography/Inline';
+import { SectionTitle } from '@/components/Typography/SectionTitle';
 
 import { rem } from '@/css';
 import technologies from '@static/technologies.json' assert { type: 'json' };
-import { ButtonGroup } from '../../components/Button/ButtonGroup';
-import { Heading } from '../../components/Typography/Heading';
-import { Paragraph } from '../../components/Typography/Paragraph';
+import { ButtonGroup } from '@/components/Button/ButtonGroup';
+import { Heading } from '@/components/Typography/Heading';
+import { Paragraph } from '@/components/Typography/Paragraph';
 
 export const Technologies: React.FC = () => {
   return (
